updated consumables value to float. cosumables still restor correct mana/health. updated consumables tool tip to account for change. updated some wording for tooltips to be clearer. added some extra info to weapon tool tips

+++ Assets/Scripts/ScriptableObjs/SOConsumables.cs
@@ -12,5 +12,5 @@ public class SOConsumables : SOItems
 	}
 
 	[Header("Percentage Value")]
-	public int consumablePercentage;
+	public float consumablePercentage;
 }
